Move test_avoid_uih_one_output to WantsOutputs containing mod
