+++ payjoin/src/core/receive/v1/mod.rs
@@ -715,31 +715,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn test_avoid_uih_one_output() {
-        let proposal = unchecked_proposal_from_test_vector();
-        let proposal_psbt = Psbt::from_str(RECEIVER_INPUT_CONTRIBUTION).unwrap();
-        let input = InputPair::new(
-            proposal_psbt.unsigned_tx.input[1].clone(),
-            proposal_psbt.inputs[1].clone(),
-            None,
-        )
-        .unwrap();
-        let input_iter = [input].into_iter();
-        let mut payjoin = wants_outputs_from_test_vector(proposal)
-            .commit_outputs()
-            .contribute_inputs(input_iter.clone())
-            .expect("Failed to contribute inputs");
-
-        payjoin.payjoin_psbt.outputs.pop();
-        let avoid_uih = payjoin.avoid_uih(input_iter);
-        assert_eq!(
-            avoid_uih.unwrap_err(),
-            SelectionError::from(InternalSelectionError::UnsupportedOutputLength),
-            "Payjoin below minimum allowed outputs for avoid uih and should error"
-        );
-    }
-
     /// Add keypath data to psbt to be prepared and verify it is excluded from the final PSBT
     /// See: <https://github.com/bitcoin/bips/blob/master/bip-0078.mediawiki#senders-payjoin-proposal-checklist>
     #[test]
